+++ lib/simple_flow.rb
@@ -16,6 +16,7 @@ require 'logger'
 
 require_relative 'simple_flow/version'
 require_relative 'simple_flow/result'
+require_relative 'simple_flow/dependency_graph'
 require_relative 'simple_flow/pipeline'
 require_relative 'simple_flow/middleware'
 require_relative 'simple_flow/step_tracker'